convert struct enum to enum class

+++ src/brickred/internal_logger.cc
@@ -7,8 +7,8 @@ namespace brickred {
 BRICKRED_PRECREATED_SINGLETON_IMPL(InternalLogger)
 
 static void defaultLogFunc(int level, const char *format, va_list args) {
-    if (level < InternalLogger::LogLevel::MIN ||
-        level >= InternalLogger::LogLevel::MAX) {
+    if (level < (int)InternalLogger::LogLevel::MIN ||
+        level >= (int)InternalLogger::LogLevel::MAX) {
         return;
     }
